v4l2-ioctl: Do not warn on device-specific pixelformat
patch from mendel-imx_8m
commit d4d0fc5c06dc2fbdf3e56905940ea15033d7843d

The v4l2 core now sets the v4l2_fmtdesc->description string based on
numeric pixelformat so that drivers no longer have to contain description strings.
This happens even if the driver already filled the description and
prints an ugly warning if the pixelformat is unknown. Reduce this to
pr_debug if the driver provided a description string.

Change-Id: I32e6587a444dd81044d8db9c33caab673bbcc0bb